Player Move Left/Right Added, Level Added, Tile Map Collision Detection Added.

+++ KnightAdventure/LTexture.cpp
@@ -44,7 +44,7 @@ void LTexture::setAlpha(Uint8 alpha) {
 	SDL_SetTextureAlphaMod(mTexture, alpha);
 }
 
-void LTexture::render(int x, int y, SDL_Rect* clip, double angle, SDL_Point* center, SDL_RendererFlip flip) {
+void LTexture::render(float x, float y, SDL_Rect* clip, double angle, SDL_Point* center, SDL_RendererFlip flip) {
 	SDL_Rect renderQuad = { x, y, mWidth, mHeight };
 	if (clip != NULL) {
 		renderQuad.w = clip->w;
